Remove GetDeviceList API call
The pts-bot was calling GetDeviceList() to query and validate the
virtual port created with wine/windows registry. However, based on
PTS API documentation it seems to check only for physical devices
and meant for native Windows env, causing pts-bot to terminate
the execution in the latest PTS versions. PTS is still able to talk
to the chip by calling SetPTSDevice() API with the virtual port
number. Hence, removing this redundant API call, which is
only meaningful for devices directly connected via USB or COM
on native Windows setup.

Tested on PTS V8.0.3 and V8.9.0